Now using setObject for binding sql params.
When reading or writing data, the appropriate type mapping is now
entirely handled by the driver used. May add support for providing a
custom type map in the future.

Also added possibility to retrieve a string representation of the last
query executed by the DatabaseAccessor. The value depends on how the
driver implements the Statement's toString method.